Updated file structure around navigation
