+++ src/Hypzeh.Web/ClientApp/components/App/components/layout/index.jsx
@@ -1,3 +1,3 @@
 export { default as Main } from './Main';
 export { default as Navbar } from './Navbar';
-export { default as Page } from './Page';
+export { default as Panel } from './Panel';
