+++ src/Hypzeh.Web/ClientApp/components/App/components/sections/ns/Home.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Page } from '../../layout';
+import { Page } from '../../shared';
 
 const Home = () => (
   <Page title="Home">
