--- src/Hypzeh.Web/ClientApp/components/App/components/sections/ns/utils/routing.js
@@ -1,18 +1,0 @@
-import { lazy } from 'react';
-
-export const defaultPath = '/';
-
-const routing = [
-  {
-    path: `${defaultPath}`,
-    title: 'Home',
-    component: lazy(() => import(/* webpackChunkName: "ns-home" */ '../Home')),
-  },
-  {
-    path: `${defaultPath}about`,
-    title: 'About',
-    component: lazy(() => import(/* webpackChunkName: "ns-about" */ '../About')),
-  },
-];
-
-export default routing;
