--- src/Hypzeh.Web/ClientApp/components/App/components/sections/smallify/utils/routing.js
@@ -1,18 +1,0 @@
-import { lazy } from 'react';
-
-export const defaultPath = '/smallify';
-
-const routing = [
-  {
-    path: `${defaultPath}`,
-    title: 'Smallify',
-    component: lazy(() => import(/* webpackChunkName: "smallify-overview" */ '../Overview')),
-  },
-  {
-    path: `${defaultPath}/screenshots`,
-    title: 'Screenshots',
-    component: lazy(() => import(/* webpackChunkName: "smallify-screenshots" */ '../Screenshots')),
-  },
-];
-
-export default routing;
