+++ src/Hypzeh.Web/ClientApp/components/App/components/shared/index.jsx
@@ -1,2 +1,3 @@
 export { default as Loader } from './Loader';
 export { default as Scroller } from './Scroller';
+export { default as Page } from './Page';
