+++ src/Hypzeh.Web/ClientApp/components/App/contexts/navigation/navigation-state.js
@@ -1,7 +1,5 @@
-import ns from '../../utils/navigation/sections/ns';
+import ns from '../../utils/navigation/ns';
 
-const state = {
-  ...ns,
-};
+const state = { ...ns };
 
 export default state;
